Several changes to service layer

+++ src/main/java/mx/kinich49/expensetracker/models/database/MonthlyBudgetCategory.java
@@ -26,6 +26,6 @@ public class MonthlyBudgetCategory {
     private Category category;
 
     @Column
-    private long monthlyLimit;
+    private int monthlyLimit;
 
 }
